Desarrollo

See merge request Genesig/Main/plugins/plugins/nomenclador!22

+++ server/DBConections/Postgree_9_1.php
@@ -48,7 +48,7 @@ class Postgree_9_1 extends DBConn
     }
     public function query($query)
     {
-        $result = pg_query($query);
+        $result = pg_query($this->dbHandler,$query);
         $this->lastResult = $result;
         if(!$result || pg_result_error($result))
             return false;
@@ -67,7 +67,7 @@ class Postgree_9_1 extends DBConn
                 $haveDefault = isset($value['default']);
                 $createStr .= "\"{$value['id']}\" {$value['type']}";
                 if($haveDefault){
-                    $createStr .= " DEFAULT '{$value['default']}'";
+                    $createStr .= " DEFAULT {$value['default']}";
                 }
                 $createStr.=',';
             }
@@ -84,8 +84,10 @@ class Postgree_9_1 extends DBConn
 
     }
 
-    public function insertData($tableName, $fieldsOrder, $schema, $data, $returning)
+    public function insertData($tableName, $fieldsOrder, $schema, $data, $returning, $isMultiEnum = false)
     {
+        if(!is_array($data) || !count($data))
+            return true;
         $f = '(';
         foreach ($fieldsOrder as $value) {
             $f.="\"$value\",";
@@ -95,14 +97,26 @@ class Postgree_9_1 extends DBConn
 
         $createStr = "INSERT INTO \"$schema\".\"$tableName\" $f VALUES";
         
-        foreach ($data as $values){
-            
-            $createStr.="(";
-            foreach ($values as $value){                
-                $createStr.=$value.",";
+        foreach ($data as $key => $values){
+
+            if(!$isMultiEnum) {
+                $createStr.="(";
+                foreach ($fieldsOrder as $field) {
+                    $createStr .= $values[$field] . ",";
+
+                }
+                $createStr = substr($createStr,0,-1);
+                $createStr .="),";
             }
-            $createStr = substr($createStr,0,-1);
-            $createStr .="),";
+            else{
+                foreach ($values as $v) {
+                    $createStr.="(";
+                    $createStr = "$createStr $key, $v";
+                    $createStr .="),";
+                }
+            }
+
+
         }
         $createStr = substr($createStr,0,-1);
         if($returning) {
@@ -154,13 +168,15 @@ class Postgree_9_1 extends DBConn
 
     public function updateData($tableName, $schema, $data)
     {
+        if(count($data)===0)
+            return true;
         $query = "";
 
         foreach ($data as $record){
             $query.="UPDATE \"$schema\".\"$tableName\" SET ";
     
             $primaryKeyVal = null;
-            $fields = array();
+            $fields = array(PrimaryKey::ID);
             foreach ($record as $fieldName => $fieldValue){
                 if ($fieldName == PrimaryKey::ID) {
                     $primaryKeyVal = $fieldValue;
@@ -178,14 +194,19 @@ class Postgree_9_1 extends DBConn
 
     }
 
-    public function deleteData($tableName, $schema, $data, $primaryField)
+    public function deleteData($tableName, $schema, $data, $primaryField, $recordPK=null,$deleteFromKey=false)
     {
+        if(count($data)===0)
+            return true;
         if(is_null($primaryField))
             $primaryField = PrimaryKey::ID;
+        if(is_null($recordPK))
+            $recordPK = $primaryField;
 
         $query = "DELETE FROM \"$schema\".\"$tableName\" WHERE ";
-        foreach ($data as $record){
-            $query.="\"".$primaryField."\""."=".$record[$primaryField]." or ";
+        foreach ($data as $key => $record){
+            $toDel = $deleteFromKey ? $key : $record[$recordPK];
+            $query.="\"".$primaryField."\""."=".$toDel." or ";
         }
 
         $query = substr($query,0,-3);
@@ -346,22 +367,27 @@ class Postgree_9_1 extends DBConn
         return $query;
     }
 
-    public function getEnumData($schema, $baseName, $select, $from, $where, $offset=null, $limit=null, $idRow=null)
+    public function getEnumData($schema, $baseName, $subQName, $selectSubq, $fromSub,$select,$from, $whereSubq, $offset=null, $limit=null, $idRow=null )
     {
         $search_path ="set search_path = $schema ";
-        $query = "$select $from";
+        $query = "$selectSubq $fromSub";
         
         if($idRow){
             $query= "$query WHERE $baseName.".PrimaryKey::ID."='$idRow' ";
         }
-        else if($where){
-            $query= "$query $where ";
+        else if($whereSubq){
+            $query= "$query $whereSubq ";
         }
+        $query ="$query order by $baseName asc ";
         if($offset){
             $query="$query OFFSET $offset";
         }
         if($limit){
             $query="$query LIMIT $limit";
+        }
+
+        if(is_string($from)){
+            $query = " $select from ($query) as $subQName $from";
         }
         $query = "$search_path ; $query ;";
         return $this->query($query);
@@ -406,9 +432,11 @@ class Postgree_9_1 extends DBConn
     public function continueFromMultiSelect( $enumSchema, $enumTableName,$currentTableName, $multiTableName, $query, $baseName)
     {
         $primaryKey = PrimaryKey::ID;
+        $multiTable = "\"$enumSchema\".\"$multiTableName\"";
+        $baseTable = "\"$enumSchema\".\"$enumTableName\" ";
         //El nombre de la tabla de un nomenclador es el identificador de un nomenclador.
-        $query = "$query inner join \"$enumSchema\".\"$multiTableName\" on ($baseName.$primaryKey=\"$currentTableName\") ";
-        $query = "$query inner join \"$enumSchema\".\"$enumTableName\" on ( \"$enumTableName\" = \"$enumSchema\".\"$enumTableName\".$primaryKey) ";
+        $query = "$query left join $multiTable on ($baseName.$primaryKey= $multiTable.\"$currentTableName\") ";
+        $query = "$query left join $baseTable on ( $multiTable.\"$enumTableName\" = $baseTable.$primaryKey) ";
         return $query;
     }
